requirements.txt
Django==4.2.10
django-stubs==4.2.7
psycopg2==2.9.9
django-ace==1.32.4
mypy==1.8.0
PyYAML==6.0.1
requests==2.31.0
djangorestframework==3.15.0
djangorestframework-types==0.8.0
types-requests==2.31.0.20240311
# boto3==1.34.59
# django-semantic-admin==0.4.1